change link to instructions for launching instances
I’m working on migrating the information on how to launch your own AMI instance from the Cloud lesson to the Genomics workshop homepage. The instructions in the cloud lesson were very detailed and included other options besides AMI. I’m proposing the following structure (discussed in this PR):
A page within the Genomics Workshop Overview with instructions on how to launch and log in to an AMI instance that is identical to that used in our workshops
A page within the Cloud lesson with additional information about modifying AMI instances, choosing among different cloud platforms, etc.