Update AMI setup to use dcuser instead of ubuntu
There are two users on the DC-genomics AMI, `dcuser` and `ubuntu`, if the user logs in via ubuntu, then they may run into permissions issues when navigating the course. Moreover, the `ubuntu` user had root access, which is never a good idea. The previous instructions had a difference between macOS and windows where windows logged in via `dcuser` and macOS logged in via `ubuntu`. I've added the following:

 - modified macOS instructions to use `dcuser` and not include the .pem file
 - update the heading of the section to be "macOS/Linux" since these instructions are sufficient for any UNIX system

The only thing that is hanging now is how to instruct the users to use the `ubuntu` account should they even need it. For macOS and Linux, users can add `-i DC-AMI.pem` to their ssh command, but for windows users, they need to [generate a private PuTTY key] from their pem file. 

Additionally, while it's mentioned that the users need a password, there's no mention of where they are supposed to get the password, but I'm not sure where that can go. 

In a way, this is an extension of #76, which is an extension of https://github.com/datacarpentry/cloud-genomics/issues/78

[generate a private PuTTY key]: https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/putty.html#putty-private-key